backend/requirements.txt
fastapi
uvicorn
osmnx
networkx
numpy
scikit-learn
python-dotenv